'fix' building document with XeLaTeX
-- sorry no latex guru here --

Following error occured for me using the 
> \usepackage[T1]{fontenc}
> Command \\nobreakspace unavailable in encoding TU.

According to https://tex.stackexchange.com/questions/66949/command-nobreakspace-unavailable-when-switching-to-t1-encoding-under-xelatex i changed line 26. 
Now the document is building again. 

Of course this is just a proposal to fix the build. 
Maybe someone knows a better approach. 

I know there isn't much recent activity on this repository and some packages may be outdated. 
So if someone has a more up to date template i would appreciate.